% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_auth.r
\name{mt_auth}
\alias{mt_auth}
\title{Get MyTracker Authentication Token}
\usage{
mt_auth(url, api_user_id, secret_key)
}
\arguments{
\item{url}{Target URL}

\item{api_user_id}{API User ID}

\item{secret_key}{Secret Key}
}
\description{
Get MyTracker Authentication Token <https://tracker.my.com/docs/api/export-api/authentication>
}
