% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_download.r
\name{mt_download}
\alias{mt_download}
\title{Download and Unzip MyTracker Data}
\usage{
mt_download(download_link, file_path, overwrite = F)
}
\arguments{
\item{download_link}{Download Link}

\item{file_path}{File Path}

\item{overwrite}{If to overwrite the existing file}
}
\description{
Download and Unzip MyTracker Data
}
