% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_download.r
\name{mt_raw_download}
\alias{mt_raw_download}
\title{Download and Unzip MyTracker Data}
\usage{
mt_raw_download(download_link, file_path)
}
\arguments{
\item{download_link}{Download Link}

\item{file_path}{File Path}
}
\description{
Download and Unzip MyTracker Data
}
