% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_raw_download_link.r
\name{mt_raw_download_link}
\alias{mt_raw_download_link}
\title{Get MyTracker Raw Data Download Link}
\usage{
mt_raw_download_link(export_id, api_user_id, secret_key, nap_time = 5)
}
\arguments{
\item{export_id}{Export ID}

\item{api_user_id}{API User ID}

\item{secret_key}{Secret Key}

\item{nap_time}{Nap time before re-trying to get the download link}
}
\description{
Get MyTracker Raw Data Download Link <https://tracker.my.com/docs/api/export-api/raw/create>
}
