% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_raw_export_id.r
\name{mt_raw_export_id}
\alias{mt_raw_export_id}
\title{Get MyTracker Raw Export ID}
\usage{
mt_raw_export_id(
  event,
  selectors,
  date_from,
  date_to,
  api_user_id,
  secret_key,
  ...
)
}
\arguments{
\item{event}{Event}

\item{selectors}{Selectors}

\item{date_from}{Start date}

\item{date_to}{End date}

\item{api_user_id}{API User ID}

\item{secret_key}{Secret Key}

\item{...}{Any other query parameters}
}
\description{
Get MyTracker Raw Export ID <https://tracker.my.com/docs/api/export-api/raw/create>
}
