% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_url_build.r
\name{mt_url_build}
\alias{mt_url_build}
\title{Get MyTracker URL}
\usage{
mt_url_build(endpoint, ...)
}
\arguments{
\item{endpoint}{Target Endpoint}

\item{...}{Any query parameters}
}
\description{
Get MyTracker URL
}
